forgot to do a commit in a while but basically just fleshing out core npc; most functions have a basic completeish setup

+++ src/core/Person.java
@@ -4,9 +4,9 @@ package core;
 public class Person {
     public String name;
     public int age;
-    protected int peeMeter;
-    protected int drunkMeter;
-    protected int grubMeter;
+    protected double peeMeter;
+    protected double drunkMeter;
+    protected double grubMeter;
 
     void consume(Item item){
